Filter out additional annotations when comparing to CR
The operator will compare the annotations the CR is configured with
with the annotations of the deployed pod. If these are different, it will
re-deploy the pod. On some implementations of kubernetes 1.19, it adds
additional automatic annotations for cni.projectcalico.org/*
that the CR will not contain. If these are not filtered out
before the map slice comparison, then the operator will always
think the annotations are different resulting in a never-ending
loop of restarting the stan pods.